updated chat component structure
